fix: rename chunkMessgae.js to chunkMessage.js to correct typo
- Rename file from chunkMessgae.js to chunkMessage.js
- Update require statement in index.js to match new filename

This improves code maintainability and fixes the spelling error.
